fix c++ build
